+++ zapisy/apps/users/models.py
@@ -1,67 +1,41 @@
 import datetime
 import logging
+from typing import List, TYPE_CHECKING, Optional
 
 from django.db import models
 from django.conf import settings
-from django.apps import apps
-from django.contrib.auth.models import User, UserManager
+from django.contrib.auth.models import User
 from django.contrib.sites.models import Site
-from django.core.mail import send_mail
-from django.template import Context
-from django.template.loader import render_to_string
-from django.core.exceptions import ValidationError
+from django.db.models import QuerySet
 
-from apps.users.exceptions import NonEmployeeException, NonStudentException, NonUserException
+from apps.users.exceptions import NonUserException
 from apps.users.managers import GettersManager, T0Manager
+
+# The TYPE_CHECKING constant is always False at runtime, so the import won't be evaluated,
+# but mypy (and other type-checking tools) will evaluate the contents of that block.
+# It protects us from circular imports.
+
+if TYPE_CHECKING:
+    from apps.enrollment.courses.models.semester import Semester
+    from apps.enrollment.courses.models.course import Course
+    from apps.offer.preferences.models import Preference
 
 logger = logging.getLogger()
 
 EMPLOYEE_STATUS_CHOICES = [(0, 'aktywny'), (1, 'nieaktywny')]
+ISIM_PROGRAM_NAME = 'ISIM, dzienne I stopnia'
 
 
 class Related(models.Manager):
-    def get_queryset(self):
+    def get_queryset(self) -> QuerySet:
         return super(Related, self).get_queryset().select_related('user')
 
 
-class ExtendedUser(User):
-    is_student = models.BooleanField(default=False, verbose_name="czy student?")
-    is_employee = models.BooleanField(default=False, verbose_name="czy pracownik?")
-    is_zamawiany = models.BooleanField(default=False, verbose_name="czy zamawiany?")
-
-    objects = UserManager()
-
-    class Meta:
-        verbose_name = 'użutkownik'
-        verbose_name_plural = 'użytkownicy'
-
-
-class UserProfile(models.Model):
-    # This field is required.
-    user = models.OneToOneField(User, related_name='profile', on_delete=models.CASCADE)
-    is_student = models.BooleanField(default=False, verbose_name="czy student?")
-    is_employee = models.BooleanField(default=False, verbose_name="czy pracownik?")
-    is_zamawiany = models.BooleanField(default=False, verbose_name="czy zamawiany?")
-    preferred_language = models.CharField(
-        max_length=5,
-        choices=settings.LANGUAGES,
-        default=settings.LANGUAGES[0][0],
-        verbose_name="preferowany język Systemu Zapisów")
-
-    def clean(self):
-        super(UserProfile, self).clean()
-        if not (self.is_employee or self.is_student) or (self.is_student and self.is_employee):
-            raise ValidationError(
-                message={
-                    'integrity': ['Profil musi jedoznacznie określać rolę użytkownika w systemie']},
-            )
-
-
 class BaseUser(models.Model):
-    '''
+    """
     User abstract class. For every app user there is entry in django.auth.
     We do not inherit after User directly, because of problems with logging beckend etc.
-    '''
+    """
     receive_mass_mail_enrollment = models.BooleanField(
         default=True,
         verbose_name="otrzymuje mailem ogłoszenia Zapisów")
@@ -75,11 +49,11 @@ class BaseUser(models.Model):
 
     objects = Related()
 
-    def get_full_name(self):
+    def get_full_name(self) -> str:
         return self.user.get_full_name()
     get_full_name.short_description = 'Użytkownik'
 
-    def get_number_of_news(self):
+    def get_number_of_news(self) -> int:
         from apps.news.models import News
         if not hasattr(self, '_count_news'):
             self._count_news = News.objects.exclude(
@@ -88,7 +62,12 @@ class BaseUser(models.Model):
         return self._count_news
 
     @staticmethod
-    def get(user_id):
+    def get(user_id: int) -> User:
+        """Returns user with specified id.
+
+        Raises:
+            NonUserException: If user with specified id doesn't exist.
+        """
         try:
             user = User.objects.get(id=user_id)
         except User.DoesNotExist:
@@ -99,24 +78,25 @@ class BaseUser(models.Model):
         return user
 
     @staticmethod
-    def is_student(user):
-        return hasattr(user, "student") and user.student
+    def is_student(user: User) -> bool:
+        if user:
+            return user.groups.filter(name='students').exists()
+        return False
 
     @staticmethod
-    def is_employee(user):
-        return hasattr(user, "employee") and user.employee
+    def is_employee(user: User) -> bool:
+        if user:
+            return user.groups.filter(name='employees').exists()
+        return False
 
-    def __str__(self):
-        return self.get_full_name
+    def __str__(self) -> str:
+        return self.get_full_name()
 
     class Meta:
-        abstract = True
+        abstract: bool = True
 
 
 class Employee(BaseUser):
-    '''
-    Employee.
-    '''
 
     user = models.OneToOneField(
         User,
@@ -132,7 +112,7 @@ class Employee(BaseUser):
         verbose_name="Status")
     title = models.CharField(max_length=20, verbose_name="tytuł naukowy", null=True, blank=True)
 
-    def has_privileges_for_group(self, group_id):
+    def has_privileges_for_group(self, group_id: int) -> bool:
         """
         Method used to verify whether user is allowed to create a poll for certain group
         (== he is an admin, a teacher for this course or a teacher for this group)
@@ -148,21 +128,15 @@ class Employee(BaseUser):
                 group_id)
         return False
 
-    def get_sex(self):
-        if self.user.first_name[-1] == 'a':
-            return 'K'
-        else:
-            return 'M'
-
     @staticmethod
-    def get_actives():
+    def get_actives() -> QuerySet:
         return Employee.objects.filter(user__is_active=True).order_by('user__last_name', 'user__first_name'). extra(
             where=["(SELECT COUNT(*) FROM courses_courseentity WHERE courses_courseentity.status > 0 AND NOT courses_courseentity.deleted AND courses_courseentity.owner_id=users_employee.id)>0"]
         )
 
     @staticmethod
-    def get_list(begin='All'):
-        def next_char(begin):
+    def get_list(begin: str ='All') -> QuerySet:
+        def next_char(begin: str) -> str:
             try:
                 return chr(ord(begin) + 1)
             except ValueError:
@@ -180,38 +154,6 @@ class Employee(BaseUser):
 
         return employees
 
-    @staticmethod
-    def get_all_groups_in_semester(user_id):
-        from apps.enrollment.courses.models.group import Group
-        from apps.enrollment.courses.models.semester import Semester
-
-        user = User.objects.get(id=user_id)
-        semester = Semester.get_default_semester()
-        try:
-            employee = user.employee
-            groups = Group.objects.filter(teacher=employee, course__semester=semester)
-        except Employee.DoesNotExist:
-            logger.error(
-                'Function Employee.get_all_groups(user_id = %d) throws Employee.DoesNotExist exception.' %
-                user_id)
-            raise NonEmployeeException()
-        return groups
-
-    @staticmethod
-    def get_all_groups(user_id):
-        from apps.enrollment.courses.models.group import Group
-
-        user = User.objects.get(id=user_id)
-        try:
-            employee = user.employee
-            groups = Group.objects.filter(teacher=employee)
-        except Employee.DoesNotExist:
-            logger.error(
-                'Function Employee.get_all_groups(user_id = %d) throws Employee.DoesNotExist exception.' %
-                user_id)
-            raise NonEmployeeException()
-        return groups
-
 
     class Meta:
         verbose_name = 'pracownik'
@@ -222,14 +164,11 @@ class Employee(BaseUser):
             ("mailto_all_students", "Może wysyłać maile do wszystkich studentów"),
         )
 
-    def __str__(self):
+    def __str__(self) -> str:
         return self.user.get_full_name()
 
 
 class Student(BaseUser):
-    '''
-    Student.
-    '''
 
     user = models.OneToOneField(
         User,
@@ -257,8 +196,6 @@ class Student(BaseUser):
 
     t0 = models.DateTimeField(null=True, blank=True)
 
-    isim = models.BooleanField(default=False)
-
     ects_in_semester = models.SmallIntegerField(default=0)
 
     dyskretna_l = models.BooleanField(default=False)
@@ -266,25 +203,25 @@ class Student(BaseUser):
     algorytmy_l = models.BooleanField(default=False)
     programowanie_l = models.BooleanField(default=False)
 
-    objects = GettersManager()
+    objects: GettersManager = GettersManager()
 
-    def make_t0(self, semester=None):
-        from apps.enrollment.courses.models.semester import Semester
-        if not semester:
-            semester = Semester.get_current_semester()
-        self.t0 = semester.records_opening - self.get_t0_interval()
-
-    def is_active(self):
+    def is_active(self) -> bool:
         return self.status == 0
 
-    def get_sex(self):
-        if self.user.first_name[-1] == 'a':
-            return 'K'
-        else:
-            return 'M'
+    def is_isim(self) -> bool:
+        try:
+            return self.program == Program.objects.get(name=ISIM_PROGRAM_NAME)
+        except Program.DoesNotExist:
+            return False
 
-    def get_type_of_studies(self):
-        """ returns type of studies """
+    def consent_answered(self) -> bool:
+        return hasattr(self, 'consent')
+
+    def consent_granted(self) -> bool:
+        return self.consent_answered() and self.consent.granted
+
+    def get_type_of_studies(self) -> str:
+        """Returns type of studies."""
         semestr = {
             1: 'pierwszy',
             2: 'drugi',
@@ -301,11 +238,11 @@ class Student(BaseUser):
         return '%s, semestr %s' % (self.program, semestr)
     get_type_of_studies.short_description = 'Studia'
 
-    def participated_in_last_grades(self):
+    def participated_in_last_grades(self) -> int:
         from apps.grade.ticket_create.models.student_graded import StudentGraded
         return StudentGraded.objects.filter(student=self, semester__in=[45, 239]).count()
 
-    def get_t0_interval(self):
+    def get_t0_interval(self) -> datetime.timedelta:
         """ returns t0 for student->start of records between 10:00 and 22:00; !record_opening hour should be 00:00:00! """
         if hasattr(self, '_counted_t0'):
             return self._counted_t0
@@ -321,34 +258,7 @@ class Student(BaseUser):
             minutes=minutes + grade + 120) + datetime.timedelta(days=3)
         return self._counted_t0
 
-    def get_voted_courses(self, given_points):
-        """ returns courses which were voted with given_points by student """
-        from apps.enrollment.courses.models.semester import Semester
-        current_semester = Semester.get_default_semester()
-        from apps.offer.vote.models.single_vote import SingleVote
-        return [
-            x.course for x in SingleVote.objects.filter(
-                student=self,
-                state__semester_winter=current_semester,
-                correction=given_points).select_related('course').order_by('course__entity__name')]
-        # return map(lambda x: x.course,
-        # StudentOptions.objects.filter(course__semester__id__exact=current_semester.id).filter(student=self,
-        # records_opening_bonus_minutes=minutes).order_by('course__name'))
-
-    def get_records_history(self, default_semester=None):
-        '''
-        Returns list of ids of course s that student was enrolled for.
-        '''
-        if not default_semester:
-            from apps.enrollment.courses.models.semester import Semester
-            default_semester = Semester.get_default_semester()
-        records = self.records.exclude(
-            group__course__semester=default_semester).select_related(
-            'group', 'group__course', 'group__course__entity')
-        records_list = [x.group.course.entity.id for x in records]
-        return list(frozenset(records_list))
-
-    def get_points(self, semester=None):
+    def get_points(self, semester: 'Semester'=None) -> int:
         from apps.enrollment.courses.models.semester import Semester
         from apps.enrollment.courses.models.points import StudentPointsView
         from apps.enrollment.records.models import Record
@@ -364,7 +274,7 @@ class Student(BaseUser):
 
         return StudentPointsView.get_points_for_entities(self, records)
 
-    def get_points_with_course(self, course, semester=None):
+    def get_points_with_course(self, course: 'Course', semester: 'Semester'=None) -> int:
         from apps.enrollment.courses.models.semester import Semester
         from apps.enrollment.courses.models.points import StudentPointsView
         from apps.enrollment.records.models import Record
@@ -382,24 +292,13 @@ class Student(BaseUser):
 
         return StudentPointsView.get_points_for_entities(self, records)
 
-    def get_schedule(self, semester=None):
-        from apps.enrollment.records.models import Record
-        from apps.enrollment.courses.models.semester import Semester
-
-        if not semester:
-            semester = Semester.get_current_semester()
-        return Record.objects.filter(status=1, group__course__semester=semester, student=self)\
-            .select_related('group', 'group__course', 'group__course__type')\
-            .prefetch_related('group__term', 'group__term__classrooms')\
-            .order_by('group__course__entity__name')
-
     @classmethod
-    def get_active_students(cls):
+    def get_active_students(cls) -> QuerySet:
         return cls.objects.filter(status=0)
 
     @staticmethod
-    def get_list(begin='All', restrict_list_consent=True):
-        def next_char(begin):
+    def get_list(begin: str='All', restrict_list_consent: Optional[bool]=True) -> QuerySet:
+        def next_char(begin: str) -> str:
             try:
                 return chr(ord(begin) + 1)
             except ValueError:
@@ -418,87 +317,20 @@ class Student(BaseUser):
             return qs.filter(user__last_name__range=(begin, end)).\
                 select_related().order_by('user__last_name', 'user__first_name')
 
-    @staticmethod
-    def get_all_groups(student):
-        try:
-            groups = [x.group for x in student.records.filter(status="1").
-                      select_related('group', 'group__teacher',
-                                     'group__course__semester',
-                                     'group__course__term')]
-        except Student.DoesNotExist:
-            logger.error('Function Student.get_all_groups(student = %d)' +
-                         'throws Student.DoesNotExist exception.' % student.pk)
-            raise NonStudentException()
-        return groups
-
-    def records_set_locked(self, locked):
+    def records_set_locked(self, locked: bool) -> None:
         self.block = locked
         self.save()
 
-    @staticmethod
-    def get_zamawiany(user_id):
-        try:
-            user = User.objects.get(id=user_id)
-            student = Student.objects.get(user=user)
-            zamawiany = StudiaZamawiane.objects.get(student=student)
-            return zamawiany
-        except (User.DoesNotExist, Student.DoesNotExist, StudiaZamawiane.DoesNotExist):
-            return None
-
-    def zamawiany(self):
-        return StudiaZamawiane.objects.get(student=self)
-
-    def zamawiany2012(self):
-        return StudiaZamawiane2012.objects.get(student=self)
-
-    # TODO: to NIE MA być pole statyczne - najlepiej zrobić mapę (pole statyczne)
-    is_zamawiany_cache = None
-
-    def is_zamawiany(self):
-        #        return self.zamawiane <> None
-
-        if not (self.is_zamawiany_cache is None):
-            return self.is_zamawiany_cache
-        try:
-            self.zamawiany()
-            self.is_zamawiany_cache = True
-        except StudiaZamawiane.DoesNotExist:
-            self.is_zamawiany_cache = False
-        return self.is_zamawiany_cache
-
-    is_zamawiany_cache2012 = None
-
-    def is_zamawiany2012(self):
-
-        #        return self.zamawiane2012 <> None
-        if not (self.is_zamawiany_cache2012 is None):
-            return self.is_zamawiany_cache2012
-        try:
-            self.zamawiany2012()
-            self.is_zamawiany_cache2012 = True
-        except StudiaZamawiane2012.DoesNotExist:
-            self.is_zamawiany_cache2012 = False
-        return self.is_zamawiany_cache2012
-
-    def is_first_year_student(self):
+    def is_first_year_student(self) -> bool:
         return (self.semestr in [1, 2]) and (self.program.id in [0, 2])
 
-    def is_fresh_student(self):
-        return True
-
-    def consent_answered(self):
-        return hasattr(self, 'consent')
-
-    def consent_granted(self):
-        return self.consent_answered() and self.consent.granted
-
     class Meta:
-        verbose_name = 'student'
-        verbose_name_plural = 'studenci'
-        app_label = 'users'
-        ordering = ['user__last_name', 'user__first_name']
+        verbose_name: str = 'student'
+        verbose_name_plural: str = 'studenci'
+        app_label: str = 'users'
+        ordering: List[str] = ['user__last_name', 'user__first_name']
 
-    def __str__(self):
+    def __str__(self) -> str:
         return self.user.get_full_name()
 
 
@@ -513,208 +345,11 @@ class Program(models.Model):
         on_delete=models.CASCADE)
 
     class Meta:
-        verbose_name = 'Program studiów'
-        verbose_name_plural = 'Programy studiów'
+        verbose_name: str = 'Program studiów'
+        verbose_name_plural: str = 'Programy studiów'
 
-    def __str__(self):
+    def __str__(self) -> str:
         return self.name
-
-
-class ZamawianeAbstract(models.Model):
-
-    points = models.FloatField(verbose_name='Punkty', null=True, blank=True)
-    comments = models.TextField(verbose_name='Uwagi', blank=True, null=True)
-    bank_account = models.CharField(
-        max_length=40,
-        null=True,
-        blank=True,
-        verbose_name="Numer konta bankowego")
-
-    class Meta:
-        abstract = True
-
-    def clean(self):
-        self.bank_account = self.bank_account.upper().replace(' ', '')
-        if not self.bank_account[:2].isalpha():
-            self.bank_account = 'PL' + self.bank_account
-        if not self.check_iban(self.bank_account):
-            raise ValidationError('Podany numer konta nie jest poprawny')
-
-    @staticmethod
-    def _normalize_char(c):
-        if c.isalpha():
-            return str(ord(c.lower()) - ord('a') + 10)
-        return c
-
-    @classmethod
-    def check_iban(cls, number):
-        """Checks if given number is valid IBAN"""
-        number = number.replace(' ', '')
-        if number == 'PL' or number == '' or number is None:
-            return True
-        lengths = {'pl': 28}
-        if not number.isalnum():
-            return False
-        country_code = number[:2].lower()
-        if not country_code.isalpha():
-            number = 'pl' + number
-            country_code = 'pl'
-        valid_length = lengths.get(country_code)
-        if valid_length is not None:
-            if len(number) != valid_length:
-                return False
-        code = int(''.join(map(cls._normalize_char, number[4:] + number[:4])))
-        return code % 97 == 1
-
-
-class StudiaZamawiane(ZamawianeAbstract):
-    """
-        Model przechowuje dodatkowe informacje o studentach zamawianych
-    """
-
-    student = models.OneToOneField(
-        Student,
-        related_name='zamawiane',
-        verbose_name='Student',
-        on_delete=models.CASCADE)
-
-    def __str__(self):
-        return 'Student zamawiany: ' + str(self.student)
-
-    def save(self, *args, **kwargs):
-        try:
-            old_sz = StudiaZamawiane.objects.get(id=self.id)
-            if self.bank_account != old_sz.bank_account and not (
-                    self.bank_account.lower() == 'pl' and old_sz.bank_account == ''):
-                current_site = Site.objects.get_current()
-                site_name, domain = current_site.name, current_site.domain
-                subject = '[Fereol] Zmiana numeru konta bankowego'
-                subject_employee = 'Zmiana numeru konta %s -> %s' % (
-                    self.student.matricula, self.bank_account and self.bank_account or '')
-                c = {
-                    'site_domain': domain,
-                    'site_name': site_name.replace('\n', ''),
-                    'student': self.student,
-                    'old_account': old_sz.bank_account and old_sz.bank_account or '',
-                    'new_account': self.bank_account and self.bank_account or '',
-                }
-                message_user = render_to_string('users/bank_account_change_email.html', c)
-                message_employee = render_to_string(
-                    'users/bank_account_change_email_employee.html', c)
-
-                emails = [x['email'] for x in StudiaZamawianeMaileOpiekunow.objects.values()]
-
-                send_mail(subject, message_user, None, [self.student.user.email])
-                send_mail(subject_employee, message_employee, None, emails)
-                logger.info(
-                    'User_id %s student_id %s has changed his bank_account to \'%s\'' %
-                    (self.student.user.id, self.student.id, self.bank_account))
-        except BaseException:
-            pass
-        if self.bank_account == '':
-            self.bank_account = None
-        super(StudiaZamawiane, self).save(*args, **kwargs)
-
-    class Meta:
-        verbose_name = 'Studia zamawiane2009'
-        verbose_name_plural = 'Studia zamawiane2009'
-
-
-class StudiaZamawiane2012(ZamawianeAbstract):
-    """
-        Model przechowuje dodatkowe informacje o studentach zamawianych
-    """
-
-    student = models.OneToOneField(
-        Student,
-        related_name='zamawiane2012',
-        verbose_name='Student',
-        on_delete=models.CASCADE)
-
-    def __str__(self):
-        return 'Student zamawiany: ' + str(self.student)
-
-    def save(self, *args, **kwargs):
-        try:
-            old_sz = StudiaZamawiane2012.objects.get(id=self.id)
-            if self.bank_account != old_sz.bank_account and not (
-                    self.bank_account.lower() == 'pl' and old_sz.bank_account == ''):
-                current_site = Site.objects.get_current()
-                site_name, domain = current_site.name, current_site.domain
-                subject = '[Fereol] Zmiana numeru konta bankowego'
-                subject_employee = 'Zmiana numeru konta %s -> %s' % (
-                    self.student.matricula, self.bank_account and self.bank_account or '')
-                c = {
-                    'site_domain': domain,
-                    'site_name': site_name.replace('\n', ''),
-                    'student': self.student,
-                    'old_account': old_sz.bank_account and old_sz.bank_account or '',
-                    'new_account': self.bank_account and self.bank_account or '',
-                }
-                message_user = render_to_string('users/bank_account_change_email.html', c)
-                message_employee = render_to_string(
-                    'users/bank_account_change_email_employee.html', c)
-
-                emails = [x['email'] for x in StudiaZamawianeMaileOpiekunow.objects.values()]
-
-                send_mail(subject, message_user, None, [self.student.user.email])
-                send_mail(subject_employee, message_employee, None, emails)
-                logger.info(
-                    'User_id %s student_id %s has changed his bank_account to \'%s\'' %
-                    (self.student.user.id, self.student.id, self.bank_account))
-        except BaseException:
-            pass
-        if self.bank_account == '':
-            self.bank_account = None
-        super(StudiaZamawiane2012, self).save(*args, **kwargs)
-
-    class Meta:
-        verbose_name = 'Studia zamawiane2012'
-        verbose_name_plural = 'Studia zamawiane2012'
-
-
-class StudiaZamawianeMaileOpiekunow(models.Model):
-    """
-        Model przechowuje maile, na które są wysyłane maile o zmianie numeru konta bankowego studentów zamawianych
-    """
-    email = models.CharField(max_length=100)
-
-    class Meta:
-        verbose_name = 'Studia zamawiane - opiekunowie'
-        verbose_name_plural = 'Studia zamawiane - opiekunowie'
-
-    def __str__(self):
-        return self.email
-
-
-"""
-CREATE OR REPLACE VIEW users_courses AS
- SELECT cc.semester_id, au.id AS student_id, cc.id AS course_id, COALESCE(
-        CASE
-            WHEN au.numeryczna_l AND cc.numeryczna_l OR au.dyskretna_l AND cc.dyskretna_l THEN ( SELECT cp.value
-               FROM courses_pointsofcourses cp
-              WHERE cp.course_id = cc.id AND cp.program_id = 1)
-            ELSE ( SELECT cp.value
-               FROM courses_pointsofcourses cp
-              WHERE cp.course_id = cc.id AND cp.program_id = au.program_id)
-        END::integer, (( SELECT cpe.value
-           FROM courses_pointsofcourseentities cpe
-          WHERE cpe.entity_id = cc.entity_id))::integer, 0) AS value, ( SELECT count(*) AS count
-           FROM records_record rr
-      LEFT JOIN courses_group cg ON rr.group_id = cg.id
-     WHERE cg.course_id = cc.id AND rr.status::integer = 1 AND rr.student_id = au.id) AS groups
-   FROM users_student au, courses_course cc;
-"""
-
-
-# definition of UserProfile from above
-# ...
-
-def create_user_profile(sender, instance, created, **kwargs):
-    if created:
-        UserProfile.objects.create(user=instance)
-
-#post_save.connect(create_user_profile, sender=User)
 
 
 class OpeningTimesView(models.Model):